+++ jPomiar/tests/testthat.R
@@ -1,0 +1,3 @@
+library(testthat)
+library(jPomiar)
+test_check("jPomiar")